fixed out of kid space bug. improved runall script.

+++ runall.sh
@@ -1,6 +1,22 @@
-
+if [ $# -gt 0 ]; then
+	postfix=$1
+	if [ $# -gt 1 ]; then
+		skip=$2
+		((skip=skip-1))
+	else
+		skip=0
+	fi
+else
+	postfix=NA
+	skip=0
+fi
 
 while read line; do
+	if [ $skip -gt 0 ]; then
+		((skip=skip-1))
+		continue
+	fi
+	echo $line
 	if [ -d mutation_out ]; then
 		rm -rf mutation_out
 	fi
@@ -8,12 +24,12 @@ while read line; do
 		rm -f gstats.xml
 	fi
 	./Controller $line/config.cfg
-	if [ -d $line/mutation_out ]; then
-		rm -rf $line/mutation_out
+	if [ -d $line/mutation_out_$postfix ]; then
+		rm -rf $line/mutation_out_$postfix
 	fi
-	if [ -f $line/gstats.xml ]; then
-		rm -f $line/gstats.xml
+	if [ -f $line/gstats_$postfix.xml ]; then
+		rm -f $line/gstats_$postfix.xml
 	fi
-	cp -r mutation_out $line/mutation_out
-	cp gstats.xml $line/gstats.xml
+	cp -r mutation_out $line/mutation_out_$postfix
+	cp gstats.xml $line/gstats_$postfix.xml
 done < subjectList.txt
